+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 * Fix getting monitor resolution for fake fullscreen
 * Save clock_id per conditional variable
 * sem_timedwait converts timeout relative to real time
+* Forked processes are set to native
 
 ## [1.4.0] - 2020-06-19
 ### Added